% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squaredv2.R
\name{squaredvs}
\alias{squaredvs}
\title{Add together two numbers}
\usage{
squaredvs(x)
}
\arguments{
\item{x}{A number to be squared}
}
\value{
the square of a number
}
\description{
Add together two numbers
}
\examples{
squares(1:3)
squares(3)
}
